+++ cmd/agentsview/pg.go
@@ -5,6 +5,8 @@ import (
 	"fmt"
 
 	"go.kenn.io/agentsview/internal/config"
+	"go.kenn.io/agentsview/internal/friction/filing"
+	"go.kenn.io/agentsview/internal/kata"
 	"go.kenn.io/agentsview/internal/postgres"
 	"go.kenn.io/agentsview/internal/server"
 	"go.kenn.io/agentsview/internal/storage"
@@ -48,6 +50,9 @@ func (pgReplica) serveOptions(
 	if rawSyncOption != nil {
 		opts = append(opts, rawSyncOption)
 	}
+	opts = append(opts, server.WithKataConn(kata.NewConn(kata.ConfigFrom(
+		appCfg.Kata, filing.EligibleHost(true, false),
+	))))
 	frictionExcl := serialExclusive()
 	frictionRunner, waitFriction := startFrictionReview(ctx, appCfg, pgStore, frictionExcl)
 	if frictionRunner != nil {