+++ internal/server/huma_route_groups.go
@@ -16,6 +16,7 @@ func (s *Server) registerTypedAPIRoutes() {
 	s.registerUsageRoutes()
 	s.registerInsightsRoutes()
 	s.registerFrictionRoutes()
+	s.registerKataRoutes()
 	s.registerSearchRoutes()
 	s.registerRecallRoutes()
 	s.describeTransferRoutes()